refactor: consolidate OAuth scope utilities
Merge `utils/oauth-scopes.ts` into `auth/oauth/scopes.ts` so OAuth-specific
helpers live next to the scope constants/`filterScopeByRole` they belong with.
The two files had no symbol overlap — `auth/oauth/scopes.ts` owned
constants + role filtering, `utils/oauth-scopes.ts` owned auth_data
serialization helpers — so this is a pure move with no behavior changes.

Importers updated to point at `auth/oauth/scopes`:
- packages/server/src/utils/oauth-connection-state.ts
- packages/server/src/connect/routes.ts
- packages/server/src/auth/social-login-provisioning.ts
- packages/server/src/tools/admin/helpers/connection-helpers.ts

The existing test (oauth-scope-filter.test.ts) already imported from the
canonical path and is unchanged.
